Refactor tenpy.linalg.misc
Move functions to tenpy.linalg.tools
The old file was a temporary solution, and this change
was intended from the beginning
