+++ tenpy/tools/string.py
@@ -1,8 +1,11 @@
 """Tools for handling strings."""
-
 # Copyright (C) TeNPy Developers, GNU GPLv3
 
-__all__ = ['is_non_string_iterable', 'vert_join', 'to_mathematica_lists', 'format_like_list']
+from typing import Sequence
+import numpy as np
+
+__all__ = ['is_non_string_iterable', 'vert_join', 'to_mathematica_lists', 'format_like_list',
+           'join_as_many_as_possible']
 
 
 def is_non_string_iterable(x):
@@ -77,6 +80,35 @@ def vert_join(strlist, valign='t', halign='l', delim=' '):
     # convert the created table to a single string
     res = '\n'.join([delim.join(lines) for lines in res])
     return res
+
+
+def join_as_many_as_possible(msgs: Sequence[str], separator: str, priorities: Sequence[int] = None,
+                             max_len: int = None, fill: str = '...') -> str:
+    """Like ``separator.join(msgs)`` but truncated if the result is too long.
+
+    We append the ``fill`` value to indicate that entries were omitted.
+    By default, the first entries in ``msgs`` are kept.
+    If ``priorities`` are specified, the messages are sorted according to their priority first
+    (from high to low).
+    """
+    if len(msgs) == 0:
+        return ''
+    if max_len is None or sum(len(m) for m in msgs) + len(separator) * (len(msgs) - 1) <= max_len:
+        if priorities:
+            return separator.join(msgs[n] for n in np.argsort(-np.array(priorities)))
+        return separator.join(msgs)
+
+    if priorities is None:
+        order = range(len(msgs))
+    else:
+        order = np.argsort(-np.array(priorities))
+
+    # build arrays whose n-th element represent the resulting length if n+1 msgs are kept
+    cum_lengths = np.cumsum([len(msgs[n]) for n in order])
+    candidate_lengths = cum_lengths + np.arange(1, len(msgs) + 1) * len(separator) + len(fill)
+    num_msgs = np.where(candidate_lengths > max_len)[0][0]
+
+    return separator.join([msgs[n] for n in order[:num_msgs]] + [fill])
 
 
 def to_mathematica_lists(a):
